+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/IssueApplicationContestedAboutToSubmitHandlerTest.java
@@ -14,9 +14,12 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.Schedule1OrMatrimonialAndCpList;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.TypeOfApplication;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.wrapper.ScheduleOneWrapper;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.GenerateCoverSheetService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.OnlineFormDocumentService;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.TestConstants.AUTH_TOKEN;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.caseDocument;
@@ -29,6 +32,8 @@ class IssueApplicationContestedAboutToSubmitHandlerTest {
     private IssueApplicationContestedAboutToSubmitHandler handler;
     @Mock
     private OnlineFormDocumentService service;
+    @Mock
+    private GenerateCoverSheetService generateCoverSheetService;
 
     @Test
     void testCanHandle() {
@@ -36,7 +41,7 @@ class IssueApplicationContestedAboutToSubmitHandlerTest {
     }
 
     @Test
-    void givenCase_whenIssueApplication_thenGenerateAppropriateCaseDocumentAndSetDefaultIfValueIsMissing() {
+    void givenCase_whenIssueApplication_thenGenerateAppropriateCaseDocument_andGenerateCoverSheets_andSetDefaultIfValueIsMissing() {
         FinremCaseDetails caseDetails = FinremCaseDetails.builder()
             .id(123L).data(FinremCaseData.builder().scheduleOneWrapper(ScheduleOneWrapper.builder().build()).build()).build();
         FinremCallbackRequest callbackRequest
@@ -46,6 +51,10 @@ class IssueApplicationContestedAboutToSubmitHandlerTest {
 
         GenericAboutToStartOrSubmitCallbackResponse<FinremCaseData> response = handler.handle(callbackRequest, AUTH_TOKEN);
         FinremCaseData data = response.getData();
+
+        verify(generateCoverSheetService).generateAndSetApplicantCoverSheet(callbackRequest.getCaseDetails(), AUTH_TOKEN);
+        verify(generateCoverSheetService).generateAndSetRespondentCoverSheet(callbackRequest.getCaseDetails(), AUTH_TOKEN);
+        verifyNoMoreInteractions(generateCoverSheetService);
 
         assertEquals("123", data.getDivorceCaseNumber());
         assertEquals(TypeOfApplication.MATRIMONIAL_CIVILPARTNERSHIP.getTypeOfApplication(),
